refactor: Refactor files structure 2
